requirements.txt
bottle
requests
psycopg2
SQLAlchemy
bottle_sqlalchemy
Pillow
jinja2
mechanize
